# data-mannagement/data-mannagement/api/dictmanage/v1/dict_manage_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.6.2
// - protoc             v5.29.3
// source: dictmanage/v1/dict_manage.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationDictManageCreateDict = "/api.dictmanage.v1.DictManage/CreateDict"
const OperationDictManageDeleteDictDirectory = "/api.dictmanage.v1.DictManage/DeleteDictDirectory"
const OperationDictManageDictDetail = "/api.dictmanage.v1.DictManage/DictDetail"
const OperationDictManageDictTemplateList = "/api.dictmanage.v1.DictManage/DictTemplateList"
const OperationDictManageEditDict = "/api.dictmanage.v1.DictManage/EditDict"
const OperationDictManageEditDictDirectoryDisabled = "/api.dictmanage.v1.DictManage/EditDictDirectoryDisabled"
const OperationDictManageEditDictDirectoryName = "/api.dictmanage.v1.DictManage/EditDictDirectoryName"
const OperationDictManageEditDictDisabled = "/api.dictmanage.v1.DictManage/EditDictDisabled"

type DictManageHTTPServer interface {
	// CreateDict 创建字典
	CreateDict(context.Context, *CreateDictRequest) (*CreateDictReply, error)
	// DeleteDictDirectory 删除字典目录（级联删除）
	DeleteDictDirectory(context.Context, *DeleteDictDirectoryRequest) (*emptypb.Empty, error)
	// DictDetail 字典详情(包括字典名字，目录结构)
	DictDetail(context.Context, *DictDetailRequest) (*DictDetailReplay, error)
	// DictTemplateList 字典模版列表
	DictTemplateList(context.Context, *DictTemplateListRequest) (*DictTemplateListReply, error)
	// EditDict 编辑字典(只会新增目录，默认为可用)
	EditDict(context.Context, *EditDictRequest) (*emptypb.Empty, error)
	// EditDictDirectoryDisabled 编辑字典目录是否可用
	EditDictDirectoryDisabled(context.Context, *EditDictDirectoryDisabledRequest) (*emptypb.Empty, error)
	// EditDictDirectoryName 编辑字典目录名字
	EditDictDirectoryName(context.Context, *EditDictDirectoryNameRequest) (*emptypb.Empty, error)
	// EditDictDisabled 编辑字典模版状态
	EditDictDisabled(context.Context, *EditDictDisabledRequest) (*emptypb.Empty, error)
}

func RegisterDictManageHTTPServer(s *http.Server, srv DictManageHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/data/dict/create", _DictManage_CreateDict0_HTTP_Handler(srv))
	r.POST("/v1/data/dict/edit", _DictManage_EditDict0_HTTP_Handler(srv))
	r.POST("/v1/data/dict/edit/status", _DictManage_EditDictDisabled0_HTTP_Handler(srv))
	r.GET("/v1/data/dict/list", _DictManage_DictTemplateList0_HTTP_Handler(srv))
	r.POST("/v1/data/dict/directory/edit/name", _DictManage_EditDictDirectoryName0_HTTP_Handler(srv))
	r.POST("/v1/data/dict/directory/edit/status", _DictManage_EditDictDirectoryDisabled0_HTTP_Handler(srv))
	r.DELETE("/v1/data/dict/directory/delete", _DictManage_DeleteDictDirectory0_HTTP_Handler(srv))
	r.GET("/v1/data/dict/detail", _DictManage_DictDetail0_HTTP_Handler(srv))
}

func _DictManage_CreateDict0_HTTP_Handler(srv DictManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateDictRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDictManageCreateDict)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateDict(ctx, req.(*CreateDictRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateDictReply)
		return ctx.Result(200, reply)
	}
}

func _DictManage_EditDict0_HTTP_Handler(srv DictManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditDictRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDictManageEditDict)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditDict(ctx, req.(*EditDictRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _DictManage_EditDictDisabled0_HTTP_Handler(srv DictManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditDictDisabledRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDictManageEditDictDisabled)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditDictDisabled(ctx, req.(*EditDictDisabledRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _DictManage_DictTemplateList0_HTTP_Handler(srv DictManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DictTemplateListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDictManageDictTemplateList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DictTemplateList(ctx, req.(*DictTemplateListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DictTemplateListReply)
		return ctx.Result(200, reply)
	}
}

func _DictManage_EditDictDirectoryName0_HTTP_Handler(srv DictManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditDictDirectoryNameRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDictManageEditDictDirectoryName)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditDictDirectoryName(ctx, req.(*EditDictDirectoryNameRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _DictManage_EditDictDirectoryDisabled0_HTTP_Handler(srv DictManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in EditDictDirectoryDisabledRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDictManageEditDictDirectoryDisabled)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.EditDictDirectoryDisabled(ctx, req.(*EditDictDirectoryDisabledRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _DictManage_DeleteDictDirectory0_HTTP_Handler(srv DictManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DeleteDictDirectoryRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDictManageDeleteDictDirectory)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DeleteDictDirectory(ctx, req.(*DeleteDictDirectoryRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _DictManage_DictDetail0_HTTP_Handler(srv DictManageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in DictDetailRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationDictManageDictDetail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.DictDetail(ctx, req.(*DictDetailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*DictDetailReplay)
		return ctx.Result(200, reply)
	}
}

type DictManageHTTPClient interface {
	CreateDict(ctx context.Context, req *CreateDictRequest, opts ...http.CallOption) (rsp *CreateDictReply, err error)
	DeleteDictDirectory(ctx context.Context, req *DeleteDictDirectoryRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	DictDetail(ctx context.Context, req *DictDetailRequest, opts ...http.CallOption) (rsp *DictDetailReplay, err error)
	DictTemplateList(ctx context.Context, req *DictTemplateListRequest, opts ...http.CallOption) (rsp *DictTemplateListReply, err error)
	EditDict(ctx context.Context, req *EditDictRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	EditDictDirectoryDisabled(ctx context.Context, req *EditDictDirectoryDisabledRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	EditDictDirectoryName(ctx context.Context, req *EditDictDirectoryNameRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	EditDictDisabled(ctx context.Context, req *EditDictDisabledRequest, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type DictManageHTTPClientImpl struct {
	cc *http.Client
}

func NewDictManageHTTPClient(client *http.Client) DictManageHTTPClient {
	return &DictManageHTTPClientImpl{client}
}

func (c *DictManageHTTPClientImpl) CreateDict(ctx context.Context, in *CreateDictRequest, opts ...http.CallOption) (*CreateDictReply, error) {
	var out CreateDictReply
	pattern := "/v1/data/dict/create"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationDictManageCreateDict))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *DictManageHTTPClientImpl) DeleteDictDirectory(ctx context.Context, in *DeleteDictDirectoryRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/dict/directory/delete"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDictManageDeleteDictDirectory))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *DictManageHTTPClientImpl) DictDetail(ctx context.Context, in *DictDetailRequest, opts ...http.CallOption) (*DictDetailReplay, error) {
	var out DictDetailReplay
	pattern := "/v1/data/dict/detail"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDictManageDictDetail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *DictManageHTTPClientImpl) DictTemplateList(ctx context.Context, in *DictTemplateListRequest, opts ...http.CallOption) (*DictTemplateListReply, error) {
	var out DictTemplateListReply
	pattern := "/v1/data/dict/list"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationDictManageDictTemplateList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *DictManageHTTPClientImpl) EditDict(ctx context.Context, in *EditDictRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/dict/edit"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationDictManageEditDict))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *DictManageHTTPClientImpl) EditDictDirectoryDisabled(ctx context.Context, in *EditDictDirectoryDisabledRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/dict/directory/edit/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationDictManageEditDictDirectoryDisabled))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *DictManageHTTPClientImpl) EditDictDirectoryName(ctx context.Context, in *EditDictDirectoryNameRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/dict/directory/edit/name"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationDictManageEditDictDirectoryName))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}

func (c *DictManageHTTPClientImpl) EditDictDisabled(ctx context.Context, in *EditDictDisabledRequest, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/v1/data/dict/edit/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationDictManageEditDictDisabled))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, err
}