# data-mannagement/data-mannagement/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/data/dict/create:
        post:
            tags:
                - DictManage
            description: 创建字典
            operationId: DictManage_CreateDict
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.dictmanage.v1.CreateDictRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.dictmanage.v1.CreateDictReply'
    /v1/data/dict/detail:
        get:
            tags:
                - DictManage
            description: 字典详情(包括字典名字，目录结构)
            operationId: DictManage_DictDetail
            parameters:
                - name: dictTemplateId
                  in: query
                  description: 模版字典id
                  schema:
                    type: integer
                    format: uint64
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.dictmanage.v1.DictDetailReplay'
    /v1/data/dict/directory/delete:
        delete:
            tags:
                - DictManage
            description: 删除字典目录（级联删除）
            operationId: DictManage_DeleteDictDirectory
            parameters:
                - name: dictDirectoryId
                  in: query
                  description: 字典目录id
                  schema:
                    type: integer
                    format: uint64
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/dict/directory/edit/name:
        post:
            tags:
                - DictManage
            description: 编辑字典目录名字
            operationId: DictManage_EditDictDirectoryName
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.dictmanage.v1.EditDictDirectoryNameRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/dict/directory/edit/status:
        post:
            tags:
                - DictManage
            description: 编辑字典目录是否可用
            operationId: DictManage_EditDictDirectoryDisabled
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.dictmanage.v1.EditDictDirectoryDisabledRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/dict/edit:
        post:
            tags:
                - DictManage
            description: 编辑字典(只会新增目录，默认为可用)
            operationId: DictManage_EditDict
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.dictmanage.v1.EditDictRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/dict/edit/status:
        post:
            tags:
                - DictManage
            description: 编辑字典模版状态
            operationId: DictManage_EditDictDisabled
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.dictmanage.v1.EditDictDisabledRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/data/dict/list:
        get:
            tags:
                - DictManage
            description: 字典模版列表
            operationId: DictManage_DictTemplateList
            parameters:
                - name: pageSize
                  in: query
                  description: 每页条数（必填）
                  schema:
                    type: integer
                    format: uint64
                - name: pageNumber
                  in: query
                  description: 页码数,从1开始（必填）
                  schema:
                    type: integer
                    format: uint64
                - name: sortBy
                  in: query
                  description: 排序字段（非必须）
                  schema:
                    type: string
                - name: order
                  in: query
                  description: 排序方式("asc" or "desc") （非必须）
                  schema:
                    type: string
                - name: searchDictName
                  in: query
                  description: 按照字典名称进行模糊搜索（非必须）
                  schema:
                    type: string
                - name: searchDictType
                  in: query
                  description: 按照字典类型进行搜索（非必须）tree:树形结构， plane：平面结构
                  schema:
                    type: string
                - name: searchDictDisabled
                  in: query
                  description: 按照是否禁用搜索, all：搜索全部，disabled：搜索禁用的 enabled：搜索未禁用的
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.dictmanage.v1.DictTemplateListReply'
...
components:
    schemas:
...
        api.dictmanage.v1.AddDictItem:
            type: object
            properties:
                parentId:
                    type: integer
                    description: 字典目录父id，新增时必须传
                    format: uint64
                name:
                    type: string
                    description: 名字
                level:
                    type: integer
                    description: 层级，根为0
                    format: uint64
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.AddDictItem'
                    description: 子目录
        api.dictmanage.v1.CreateDictReply:
            type: object
            properties:
                dictTemplateId:
                    type: integer
                    description: 字典模版id
                    format: uint64
        api.dictmanage.v1.CreateDictRequest:
            type: object
            properties:
                dictName:
                    type: string
                    description: 字典模版名称
                dictType:
                    type: string
                    description: tree:树形结构， plane：平面结构
                treeDict:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.CreateTreeDict'
                    description: 树形字典目录相关信息（非必填）
                planeDict:
                    $ref: '#/components/schemas/api.dictmanage.v1.CreatePlaneDict'
        api.dictmanage.v1.CreatePlaneDict:
            type: object
            properties:
                planeDict:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.CreatePlaneDict_Plane'
                    description: 平面字典目录相关信息（非必填）
        api.dictmanage.v1.CreatePlaneDict_Plane:
            type: object
            properties:
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
                name:
                    type: string
                    description: 目录名字
        api.dictmanage.v1.CreateTreeDict:
            type: object
            properties:
                name:
                    type: string
                    description: 名字
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.CreateTreeDict'
                    description: 子目录
        api.dictmanage.v1.Dict:
            type: object
            properties:
                id:
                    type: integer
                    description: 字典目录id(创建不用填)
                    format: uint64
                parentId:
                    type: integer
                    description: 字典目录父id，新增时必须传
                    format: uint64
                name:
                    type: string
                    description: 名字
                level:
                    type: integer
                    description: 层级，根为0
                    format: uint64
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
                isUsed:
                    type: boolean
                    description: 是否使用中，0否，1使用中
                children:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.Dict'
                    description: 子目录
        api.dictmanage.v1.DictDetailReplay:
            type: object
            properties:
                dictTemplate:
                    $ref: '#/components/schemas/api.dictmanage.v1.DictTemplate'
                treeDict:
                    $ref: '#/components/schemas/api.dictmanage.v1.TreeDict'
                planeDict:
                    $ref: '#/components/schemas/api.dictmanage.v1.PlaneDict'
        api.dictmanage.v1.DictTemplate:
            type: object
            properties:
                dictTemplateId:
                    type: integer
                    description: 模版字典id
                    format: uint64
                dictName:
                    type: string
                    description: 字典名称
                dictType:
                    type: string
                    description: 字典类型
                dictDisabled:
                    type: boolean
                    description: 字典是否禁用, 否：0，禁用：1
                updateTime:
                    type: integer
                    description: 更新时间
                    format: int64
                creatorName:
                    type: string
                    description: 创建者
                dictIsUsed:
                    type: boolean
                    description: 是否使用中，0否，1使用中
        api.dictmanage.v1.DictTemplateListReply:
            type: object
            properties:
                content:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.DictTemplate'
                    description: 字典模版列表
                totalElements:
                    type: integer
                    description: 字典模版总个数
                    format: int32
        api.dictmanage.v1.EditDictDirectoryDisabledRequest:
            type: object
            properties:
                dictDirectoryId:
                    type: integer
                    description: 字典目录id
                    format: uint64
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
        api.dictmanage.v1.EditDictDirectoryNameRequest:
            type: object
            properties:
                dictTemplateId:
                    type: integer
                    description: 模版字典id
                    format: uint64
                dictDirectoryId:
                    type: integer
                    description: 字典目录id
                    format: uint64
                dictDirectoryName:
                    type: string
                    description: 字典目录名字
        api.dictmanage.v1.EditDictDisabledRequest:
            type: object
            properties:
                dictTemplateId:
                    type: integer
                    description: 字典模版id
                    format: uint64
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
        api.dictmanage.v1.EditDictRequest:
            type: object
            properties:
                dictTemplateId:
                    type: integer
                    description: 模版字典id
                    format: uint64
                dictName:
                    type: string
                    description: 字典模版名称
                treeDict:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.AddDictItem'
                    description: 树形字典目录相关信息（非必填）
                planeDict:
                    $ref: '#/components/schemas/api.dictmanage.v1.CreatePlaneDict'
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
        api.dictmanage.v1.PlaneDict:
            type: object
            properties:
                planeDict:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.PlaneDict_Plane'
                    description: 平面字典目录相关信息（非必填）
            description: 平面字典目录
        api.dictmanage.v1.PlaneDict_Plane:
            type: object
            properties:
                id:
                    type: integer
                    description: 字典目录id(创建不用填)
                    format: uint64
                isDisabled:
                    type: boolean
                    description: 是否禁用, 否：0，禁用：1，默认0
                name:
                    type: string
                    description: 目录名字
                isUsed:
                    type: boolean
                    description: 是否使用中，0否，1使用中
        api.dictmanage.v1.TreeDict:
            type: object
            properties:
                treeDict:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.dictmanage.v1.Dict'
                    description: 树形字典目录相关信息（非必填）
            description: 树形字典目录
...